refactor(utils): Phase 3 - Move HTTP utilities to http/ subdirectory
Organizational refactor - no functional changes:
- src/utils/http/response-builder.ts (error response building)
- src/utils/http/error-status.ts (HTTP status mapping)
- src/utils/http/streaming-response.ts (SSE streaming)

Updated imports in 12 files and fixed relative paths.
All 199 smoke tests passing.

Part of Phase 3: High-risk utility consolidation.
