+++ src/handlers/metrics-handler.ts
@@ -1,6 +1,6 @@
 import type { Context } from 'hono'
 import type { Env } from '../types/env'
-import { createErrorResponse, ErrorCodes } from '../utils/response-builder'
+import { createErrorResponse, ErrorCodes } from '../utils/http/response-builder'
 
 /**
  * GET /metrics - Comprehensive metrics API endpoint
