+++ src/handlers/search-handlers.ts
@@ -9,7 +9,7 @@
 
 import * as externalApis from '../services/external-apis'
 import type { Env } from '../types/env'
-import { createErrorResponse, ErrorCodes } from '../utils/response-builder'
+import { createErrorResponse, ErrorCodes } from '../utils/http/response-builder'
 import { transformWorkToGoogleFormat } from '../utils/transform-work'
 
 // ============================================================================
