+++ src/routes/cache.ts
@@ -14,7 +14,7 @@
 import { OpenAPIHono } from '@hono/zod-openapi'
 import { handleCacheMetrics } from '../handlers/cache-metrics.js'
 import type { Env } from '../types/env'
-import { createErrorResponse, ErrorCodes } from '../utils/response-builder'
+import { createErrorResponse, ErrorCodes } from '../utils/http/response-builder'
 
 // DO stub interface for CacheMetricsDO
 interface CacheMetricsStub {
