+++ src/routes/websocket.ts
@@ -13,7 +13,7 @@
 import { OpenAPIHono } from '@hono/zod-openapi'
 import { checkRateLimit } from '../middleware/rate-limiter.ts'
 import type { Env } from '../types/env'
-import { createErrorResponse, ErrorCodes } from '../utils/response-builder'
+import { createErrorResponse, ErrorCodes } from '../utils/http/response-builder'
 
 // DO stub interfaces for RPC calls
 interface WebSocketConnectionStub {
